+++ scripts/vim-patch.sh
@@ -229,6 +229,10 @@ preprocess_patch() {
   LC_ALL=C sed -e 's/\( [ab]\/src\/nvim\)\/evalfunc\.c/\1\/eval\/funcs\.c/g' \
     "$file" > "$file".tmp && mv "$file".tmp "$file"
 
+  # Rename evalvars.c to eval/vars.c
+  LC_ALL=C sed -e 's/\( [ab]\/src\/nvim\)\/evalfunc\.c/\1\/eval\/vars\.c/g' \
+    "$file" > "$file".tmp && mv "$file".tmp "$file"
+
   # Rename userfunc.c to eval/userfunc.c
   LC_ALL=C sed -e 's/\( [ab]\/src\/nvim\)\/userfunc\.c/\1\/eval\/userfunc\.c/g' \
     "$file" > "$file".tmp && mv "$file".tmp "$file"
