+++ src/nvim/eval/executor.c
@@ -12,8 +12,6 @@
 # include "eval/executor.c.generated.h"
 #endif
 
-static char *e_letwrong = N_("E734: Wrong variable type for %s=");
-
 char *e_listidx = N_("E684: list index out of range: %" PRId64);
 
 /// Handle tv1 += tv2, -=, *=, /=,  %=, .=
