+++ src/nvim/eval/vars.h
@@ -1,0 +1,10 @@
+#ifndef NVIM_EVAL_VARS_H
+#define NVIM_EVAL_VARS_H
+
+#include "nvim/eval/funcs.h"  // For FunPtr
+#include "nvim/ex_cmds_defs.h"  // For exarg_T
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "eval/vars.h.generated.h"
+#endif
+#endif  // NVIM_EVAL_VARS_H
