+++ src/nvim/globals.h
@@ -941,6 +941,9 @@ EXTERN char e_loclist[] INIT(= N_("E776: No location list"));
 EXTERN char e_re_damg[] INIT(= N_("E43: Damaged match string"));
 EXTERN char e_re_corr[] INIT(= N_("E44: Corrupted regexp program"));
 EXTERN char e_readonly[] INIT(= N_("E45: 'readonly' option is set (add ! to override)"));
+EXTERN char e_letwrong[] INIT(= N_("E734: Wrong variable type for %s="));
+EXTERN char e_illvar[] INIT(= N_("E461: Illegal variable name: %s"));
+EXTERN char e_cannot_mod[] INIT(= N_("E995: Cannot modify existing variable"));
 EXTERN char e_readonlyvar[] INIT(= N_("E46: Cannot change read-only variable \"%.*s\""));
 EXTERN char e_stringreq[] INIT(= N_("E928: String required"));
 EXTERN char e_dictreq[] INIT(= N_("E715: Dictionary required"));
